test(otel): complete tracestate parametric migration
