+++ manifests/cpp_kong.yml
@@ -25,7 +25,6 @@ manifest:
   tests/test_ipv6.py: missing_feature
   tests/test_library_conf.py: irrelevant
   tests/test_miscs.py: irrelevant
-  tests/test_otel_tracestate_sampling.py: irrelevant (library does not implement OpenTelemetry)
   tests/test_profiling.py: irrelevant (Kong is a proxy, not a profiling target)
   tests/test_resource_renaming.py: missing_feature
   tests/test_rum_injection.py: irrelevant
