+++ manifests/golang.yml
@@ -1382,8 +1382,6 @@ manifest:
         net-http: v1.70.1
         net-http-span-pool: v1.70.1
   tests/otel/test_tracing_otlp.py::Test_Otel_Tracing_OTLP: v2.8.0-dev
-  tests/otel/test_tracing_otlp.py::Test_Otlp_Carries_Ot: missing_feature (APMAPI-2172)
-  tests/otel/test_tracing_otlp.py::Test_Otlp_Forwards_Inherited_Ot: missing_feature (APMAPI-2172)
   tests/otel_tracing_e2e/test_e2e.py::Test_OTelLogE2E: irrelevant
   tests/otel_tracing_e2e/test_e2e.py::Test_OTelMetricE2E: irrelevant
   tests/otel_tracing_e2e/test_e2e.py::Test_OTelTracingE2E: irrelevant
@@ -1848,61 +1846,6 @@ manifest:
   tests/test_library_logs.py::Test_NoExceptions::test_dotnet: irrelevant (only for .NET)
   tests/test_library_logs.py::Test_NoExceptions::test_java_logs: irrelevant (only for Java)
   tests/test_library_logs.py::Test_NoExceptions::test_java_telemetry_logs: irrelevant (only for Java)
-  tests/test_otel_tracestate_sampling.py::Test_EmitOtOnProbabilityDecision_Rate0_01: missing_feature (APMAPI-2171)
-  tests/test_otel_tracestate_sampling.py::Test_EmitOtOnProbabilityDecision_Rate0_1: missing_feature (APMAPI-2171)
-  tests/test_otel_tracestate_sampling.py::Test_EmitOtOnProbabilityDecision_Rate0_2: missing_feature (APMAPI-2171)
-  tests/test_otel_tracestate_sampling.py::Test_EmitOtOnProbabilityDecision_Rate0_5: missing_feature (APMAPI-2171)
-  tests/test_otel_tracestate_sampling.py::Test_EmitOtOnProbabilityDecision_Rate0_99: missing_feature (APMAPI-2171)
-  tests/test_otel_tracestate_sampling.py::Test_ForceKeepClearsTh:
-    - weblog_declaration:
-        "*": missing_feature (APMAPI-2171)
-        envoy: irrelevant (not applicable to proxies)
-        haproxy: irrelevant (not applicable to proxies)
-  tests/test_otel_tracestate_sampling.py::Test_ForwardInboundOtUnchanged:
-    - weblog_declaration:
-        "*": missing_feature (APMAPI-2171)
-        envoy: irrelevant (not applicable to proxies)
-        haproxy: irrelevant (not applicable to proxies)
-  tests/test_otel_tracestate_sampling.py::Test_ForwardInboundOtUnchangedWhenDropped:
-    - weblog_declaration:
-        "*": missing_feature (APMAPI-2171)
-        envoy: irrelevant (not applicable to proxies)
-        haproxy: irrelevant (not applicable to proxies)
-  tests/test_otel_tracestate_sampling.py::Test_MalformedOtHandling:
-    - weblog_declaration:
-        "*": missing_feature (APMAPI-2171)
-        envoy: irrelevant (not applicable to proxies)
-        haproxy: irrelevant (not applicable to proxies)
-  tests/test_otel_tracestate_sampling.py::Test_PrecisionBoundaryDecision_Rate0_05:
-    - weblog_declaration:
-        "*": missing_feature (APMAPI-2171)
-        envoy: irrelevant (not applicable to proxies)
-        haproxy: irrelevant (not applicable to proxies)
-  tests/test_otel_tracestate_sampling.py::Test_PrecisionBoundaryDecision_Rate0_1:
-    - weblog_declaration:
-        "*": missing_feature (APMAPI-2171)
-        envoy: irrelevant (not applicable to proxies)
-        haproxy: irrelevant (not applicable to proxies)
-  tests/test_otel_tracestate_sampling.py::Test_PreserveDdAndOtherVendors:
-    - weblog_declaration:
-        "*": missing_feature (APMAPI-2171)
-        envoy: irrelevant (not applicable to proxies)
-        haproxy: irrelevant (not applicable to proxies)
-  tests/test_otel_tracestate_sampling.py::Test_SampledWithoutOtNotFabricated:
-    - weblog_declaration:
-        "*": missing_feature (APMAPI-2171)
-        envoy: irrelevant (not applicable to proxies)
-        haproxy: irrelevant (not applicable to proxies)
-  tests/test_otel_tracestate_sampling.py::Test_ThOnlyDoesNotFabricateRv:
-    - weblog_declaration:
-        "*": missing_feature (APMAPI-2171)
-        envoy: irrelevant (not applicable to proxies)
-        haproxy: irrelevant (not applicable to proxies)
-  tests/test_otel_tracestate_sampling.py::Test_ThOnlyDoesNotFabricateRvWhenDropped:
-    - weblog_declaration:
-        "*": missing_feature (APMAPI-2171)
-        envoy: irrelevant (not applicable to proxies)
-        haproxy: irrelevant (not applicable to proxies)
   tests/test_otlp_runtime_metrics.py::Test_OtlpRuntimeMetrics: v2.10.0-dev
   tests/test_profiling.py::Test_Profile::test_process_tags_svc: missing_feature
   tests/test_resource_renaming.py:
