+++ manifests/nodejs.yml
@@ -2198,8 +2198,6 @@ manifest:
   tests/otel/test_tracing_otlp.py::Test_Otel_Tracing_OTLP::test_unsampled_trace:
     - weblog_declaration:
         nextjs: missing_feature (AppSec/IAST force-samples traces, overriding unsampled traceparent)
-  tests/otel/test_tracing_otlp.py::Test_Otlp_Carries_Ot: missing_feature (APMAPI-2172)
-  tests/otel/test_tracing_otlp.py::Test_Otlp_Forwards_Inherited_Ot: missing_feature (APMAPI-2172)
   tests/otel_tracing_e2e/test_e2e.py::Test_OTelLogE2E: irrelevant
   tests/otel_tracing_e2e/test_e2e.py::Test_OTelMetricE2E: irrelevant
   tests/otel_tracing_e2e/test_e2e.py::Test_OTelTracingE2E: irrelevant
@@ -2853,64 +2851,6 @@ manifest:
   tests/test_library_logs.py::Test_NoExceptions::test_dotnet: irrelevant (only for .NET)
   tests/test_library_logs.py::Test_NoExceptions::test_java_logs: irrelevant (only for Java)
   tests/test_library_logs.py::Test_NoExceptions::test_java_telemetry_logs: irrelevant (only for Java)
-  tests/test_otel_tracestate_sampling.py::Test_EmitOtOnProbabilityDecision_Rate0_01: missing_feature (APMAPI-2171)
-  tests/test_otel_tracestate_sampling.py::Test_EmitOtOnProbabilityDecision_Rate0_1: missing_feature (APMAPI-2171)
-  tests/test_otel_tracestate_sampling.py::Test_EmitOtOnProbabilityDecision_Rate0_2: missing_feature (APMAPI-2171)
-  tests/test_otel_tracestate_sampling.py::Test_EmitOtOnProbabilityDecision_Rate0_5: missing_feature (APMAPI-2171)
-  tests/test_otel_tracestate_sampling.py::Test_EmitOtOnProbabilityDecision_Rate0_99: missing_feature (APMAPI-2171)
-  tests/test_otel_tracestate_sampling.py::Test_ForceKeepClearsTh:  # TODO: a lower version might be supported
-    - declaration: missing_feature (APMAPI-2171)
-      component_version: <6.8.0
-      weblog: [express4-typescript, uds-express4, express4, express5, fastify, nextjs]
-    - declaration: missing_feature (APMAPI-2171)
-      excluded_weblog: [express4-typescript, uds-express4, express4, express5, fastify, nextjs]
-  tests/test_otel_tracestate_sampling.py::Test_ForceKeepClearsTh::test_force_keep_overrides_inherited_drop_decision:
-    - weblog_declaration:
-        fastify: missing_feature (APMAPI-2171)
-        express4: missing_feature (APMAPI-2171)
-        express5: missing_feature (APMAPI-2171)
-        uds-express4: missing_feature (APMAPI-2171)
-        express4-typescript: missing_feature (APMAPI-2171)
-        nextjs: missing_feature (APMAPI-2171)
-  tests/test_otel_tracestate_sampling.py::Test_ForwardInboundOtUnchanged:  # TODO: a lower version might be supported
-    - declaration: missing_feature (APMAPI-2171)
-      component_version: <6.8.0
-      weblog: [express4-typescript, uds-express4, express4, express5, fastify, nextjs]
-    - declaration: missing_feature (APMAPI-2171)
-      excluded_weblog: [express4-typescript, uds-express4, express4, express5, fastify, nextjs]
-  tests/test_otel_tracestate_sampling.py::Test_ForwardInboundOtUnchangedWhenDropped:  # TODO: a lower version might be supported
-    - declaration: missing_feature (APMAPI-2171)
-      component_version: <6.8.0
-      weblog: [express4-typescript, uds-express4, express4, express5, fastify, nextjs]
-    - declaration: missing_feature (APMAPI-2171)
-      excluded_weblog: [express4-typescript, uds-express4, express4, express5, fastify, nextjs]
-  tests/test_otel_tracestate_sampling.py::Test_MalformedOtHandling: missing_feature (APMAPI-2171)
-  tests/test_otel_tracestate_sampling.py::Test_PrecisionBoundaryDecision_Rate0_05: missing_feature (APMAPI-2171)
-  tests/test_otel_tracestate_sampling.py::Test_PrecisionBoundaryDecision_Rate0_1: missing_feature (APMAPI-2171)
-  tests/test_otel_tracestate_sampling.py::Test_PreserveDdAndOtherVendors:  # TODO: a lower version might be supported
-    - declaration: missing_feature (APMAPI-2171)
-      component_version: <6.8.0
-      weblog: [express4-typescript, uds-express4, express4, express5, fastify, nextjs]
-    - declaration: missing_feature (APMAPI-2171)
-      excluded_weblog: [express4-typescript, uds-express4, express4, express5, fastify, nextjs]
-  tests/test_otel_tracestate_sampling.py::Test_SampledWithoutOtNotFabricated:  # TODO: a lower version might be supported
-    - declaration: missing_feature (APMAPI-2171)
-      component_version: <6.8.0
-      weblog: [express4-typescript, uds-express4, express4, express5, fastify, nextjs]
-    - declaration: missing_feature (APMAPI-2171)
-      excluded_weblog: [express4-typescript, uds-express4, express4, express5, fastify, nextjs]
-  tests/test_otel_tracestate_sampling.py::Test_ThOnlyDoesNotFabricateRv:  # TODO: a lower version might be supported
-    - declaration: missing_feature (APMAPI-2171)
-      component_version: <6.8.0
-      weblog: [express4-typescript, uds-express4, express4, express5, fastify, nextjs]
-    - declaration: missing_feature (APMAPI-2171)
-      excluded_weblog: [express4-typescript, uds-express4, express4, express5, fastify, nextjs]
-  tests/test_otel_tracestate_sampling.py::Test_ThOnlyDoesNotFabricateRvWhenDropped:  # TODO: a lower version might be supported
-    - declaration: missing_feature (APMAPI-2171)
-      component_version: <6.8.0
-      weblog: [express4-typescript, uds-express4, express4, express5, fastify, nextjs]
-    - declaration: missing_feature (APMAPI-2171)
-      excluded_weblog: [express4-typescript, uds-express4, express4, express5, fastify, nextjs]
   tests/test_otlp_runtime_metrics.py::Test_OtlpRuntimeMetrics: v5.107.0-pre
   tests/test_profiling.py::Test_Profile: *ref_5_16_0
   tests/test_profiling.py::Test_Profile::test_process_tags:  # TODO: a lower version might be supported
